HADOOP-2903.  Fix exception generated by Metrics while using pushMetric(). (girish vaitheeswaran via dhruba)
git-svn-id: https://svn.apache.org/repos/asf/hadoop/core/trunk@632186 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hadoop/metrics/util/MetricsTimeVaryingRate.java
@@ -126,7 +126,7 @@ public class MetricsTimeVaryingRate {
   public synchronized void pushMetric(final MetricsRecord mr) {
     intervalHeartBeat();
     mr.incrMetric(name + "_num_ops", getPreviousIntervalNumOps());
-    mr.incrMetric(name + "_avg_time", getPreviousIntervalAverageTime());
+    mr.incrMetric(name + "_avg_time", (int)getPreviousIntervalAverageTime());
   }
   
   /**
